+++ app/controllers/staff/organisations_controller.rb
@@ -14,13 +14,11 @@ class Staff::OrganisationsController < Staff::BaseController
 
     project_activities = FindProjectActivities.new(organisation: organisation, user: current_user).call(eager_load_parent: false)
     third_party_project_activities = FindThirdPartyProjectActivities.new(organisation: organisation, user: current_user).call(eager_load_parent: false)
-    reports = ReportPolicy::Scope.new(current_user, Report).resolve.includes(:organisation, :fund)
 
     respond_to do |format|
       format.html do
         @project_activities = project_activities.map { |activity| ActivityPresenter.new(activity) }
         @third_party_project_activities = third_party_project_activities.map { |activity| ActivityPresenter.new(activity) }
-        @reports = reports.map { |report| ReportPresenter.new(report) }
       end
       format.xml do
         @activities = case level