+++ CHANGELOG.md
@@ -235,6 +235,7 @@
 - Export Activity data in the Submission CSV
 - rename Submission to Report
 - reports can be managed in the reports section of the application
+- remove submission from the organisation page
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-13...HEAD
 [release-13]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-12...release-13